Move consistency pkg tests to consistencyv1 package
