+++ api/pkg/fjarm/consistency/consistencyv1/consistency.go
@@ -1,0 +1,1 @@
+package consistencyv1
