--- api/pkg/fjarm/consistency/v1/consistency.go
@@ -1,1 +1,0 @@
-package v1
